12_django_flight_app_auth/requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.1
Django==4.2.5
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
Pillow==10.0.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
